requirements.txt
opencv-python~=4.10.0.82
numpy~=1.26.4

pandas~=2.2.2
whisperx~=3.1.3
torch~=2.3.0
transformers~=4.39.3
requests~=2.32.3
pillow~=10.3.0
googletrans~=4.0.0-rc1